Start adding the plumbing back.
The test data adding method works now, saving the data to the new Sugar database. The milestone here is that it creates new Habits, saves them, then is able to query them back out, feed them to an Adapter which can render them into Views that the RecyclerView can display. A whole lot going on, actually.