Byte is not supported by SQLite 3 so I switched to int.
I originally chose byte just because it was smaller and I'm only using one byte's worth of bits anyway. I figured this would save memory, but it turns out that SQLite 3 works with one Integer type and stores it with a variable amount of bits so it's the best of both worlds.